This PR completes two major refactoring tasks:

Completely removes the fragile C++03 psram_allocator template boilerplate, migrating to standard allocators and heap_caps_malloc_extmem_enable(16) to gracefully route >16 byte allocations to PSRAM with fallback to Internal RAM. This resolves OOM crashes while maintaining massive memory headroom.
Modernizes all Hex effects (include/effects/hex/*.h) using clang-tidy to use range-based for-loops and modern auto conventions.
Tested successfully on Four Corners build (demo, mesmerizer, tinyled-demo, tinyled-hex) and booted cleanly on Mesmerizer.

+++ include/effects/matrix/PatternSMMetaBalls.h
@@ -39,7 +39,7 @@ class PatternSMMetaBalls : public EffectWithId<PatternSMMetaBalls>
 
     void Draw() override
     {
-        for (uint8_t a = 0; a < 5; a++)
+        for (int a = 0; a < 5; a++)
         {
             bx[a] = beatsin16(15 + a * 2, 0, MATRIX_WIDTH - 1, 0, a * 32);
             by[a] = beatsin16(18 + a * 2, 0, MATRIX_HEIGHT - 1, 0, a * 32);
@@ -49,7 +49,7 @@ class PatternSMMetaBalls : public EffectWithId<PatternSMMetaBalls>
             for (unsigned j = 0; j < MATRIX_HEIGHT - 1; j++)
             {
                 uint8_t sum = dist(i, j, bx[0], by[0]);
-                for (uint8_t a = 1; a < 5; a++)
+                for (int a = 1; a < 5; a++)
                 {
                     sum = std::add_sat<uint8_t>(sum, dist(i, j, bx[a], by[a]));
                 }
