+++ include/effects/matrix/PatternSMWalkingMachine.h
@@ -38,12 +38,12 @@ class PatternSMWalkingMachine : public EffectWithId<PatternSMWalkingMachine>
     {
         g()->Clear();
 
-        for (uint8_t i = 0; i < 7; i++)
+        for (int i = 0; i < 7; i++)
         {
             dot[i].posX = (beatsin16(4, (MATRIX_WIDTH >> 3) << 8, (MATRIX_WIDTH - (MATRIX_WIDTH >> 3) - 1) << 8, i * 8192, i * 8192)) / 255.f;
             dot[i].posY = (beatsin16(4, (MATRIX_HEIGHT >> 3) << 8, (MATRIX_HEIGHT - (MATRIX_HEIGHT >> 3) - 1) << 8, i * 4096, 16384 + i * 8192)) /  255.f;
         }
-        for (uint8_t i = 0; i < 7; i++)
+        for (int i = 0; i < 7; i++)
         {
             g()->drawSafeFilledCircleF(dot[i].posX, dot[i].posY, 4, CHSV(i * 32, 255, 255));
             g()->drawLineF(dot[i].posX, dot[i].posY, dot[(i + 1) % 7].posX, dot[(i + 1) % 7].posY, CHSV(i * 32, 255, 255),
